feat: support extra html
