+++ extensions/iceworks-ui-builder/web/src/pages/canvas/app.ts
@@ -1,7 +1,6 @@
 import { createApp } from 'ice';
 import Canvas from './index';
 
-alert(123)
 const appConfig = {
   router: {
     routes: [{ path: '/', component: Canvas }],
